+++ package.json
@@ -18,11 +18,14 @@
   },
   "homepage": "https://github.com/gektin/max-descriptions#readme",
   "dependencies": {
+    "@prisma/client": "^2.7.1",
     "axios": "^0.20.0",
     "bluebird": "^3.7.2",
     "express": "^4.17.1",
     "jest": "^26.4.2",
-    "nodemon": "^2.0.4",
-    "prisma": "^1.34.10"
+    "nodemon": "^2.0.4"
+  },
+  "devDependencies": {
+    "@prisma/cli": "^2.7.1"
   }
 }
