--- project-utils/pollinator.js
@@ -1,10 +1,0 @@
-import axios from 'axios';
-
-const seedData = {
-  names: [],
-  marketValues: [],
-  logoURLs: [],
-  helpers: [],
-  minCosts: [],
-};
-
